google_slides/requirements.txt
os
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
cairosvg